fix: refactoring and redesign
