+++ src/app/app.component.html
@@ -1,3 +1,3 @@
-<app-sidebar class="header"></app-sidebar>
+<app-sidebar></app-sidebar>
 <router-outlet></router-outlet>
