--- src/app/app.component.scss
@@ -1,8 +1,0 @@
-main {
-  margin-top: 64px;
-  padding: 16px;
-}
-.header {
-  width: 100%;
-  margin-bottom: 10px;
-}
