+++ src/app/app.routes.ts
@@ -7,9 +7,12 @@ import {isAuthenticated} from './shared/guards/is-authenticated.guard';
 
 export const routes: Routes = [
   {
-    path: 'protected',
-    component: AddBlogPageComponent,
-    canActivate: [isAuthenticated]
+    path: 'create',
+    loadComponent: () =>
+        import('./features/components/add-blog-page/add-blog-page.component').then(
+            (m) => m.AddBlogPageComponent,
+        ), // Lazy loading for standalone components
+    canActivate: [isAuthenticated],
   },
   {
     path: 'blogs-overview',
