+++ src/app/core/models/blogDetails.model.ts
