+++ src/app/core/models/blogPreview.model.ts
