+++ src/app/features/components/blog-list/blog-list.component.html
@@ -1,5 +1,4 @@
 <div class="blog-list">
-  <p>asdfasdfasdf</p>
     @if (!loading && blogs.length > 0) {
         <div class="blog-grid">
             @for (blog of blogs; track blog.id) {
@@ -11,11 +10,9 @@
             }
         </div>
     }
-
     @if (loading) {
         <mat-progress-spinner mode="indeterminate" color="primary"></mat-progress-spinner>
     }
-
     @if (!loading && blogs.length === 0) {
         <mat-card>
             <mat-card-content>No blogs found.</mat-card-content>
