+++ src/app/features/components/blog-list/blog-list.component.scss
@@ -11,54 +11,9 @@
     width: 100%;
     max-width: 1200px;
   }
-
-  .blog-card {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    height: 100%;
-    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-    border-radius: 8px;
-
-    img {
-      max-height: 200px;
-      object-fit: cover;
-      width: 100%;
-      border-top-left-radius: 8px;
-      border-top-right-radius: 8px;
-    }
-
-    mat-card-title {
-      font-size: 1.4rem;
-      font-weight: bold;
-    }
-
-    mat-card-subtitle {
-      font-size: 1rem;
-      color: rgba(0, 0, 0, 0.6);
-    }
-
-    mat-card-content {
-      flex-grow: 1;
-      font-size: 0.95rem;
-      color: rgba(0, 0, 0, 0.8);
-      padding: 8px 16px;
-    }
-
-    mat-card-actions {
-      display: flex;
-      justify-content: space-between;
-      align-items: center;
-      padding: 8px 16px;
-
-      button {
-        font-weight: bold;
-      }
-
-      .meta-info {
-        font-size: 0.85rem;
-        color: rgba(0, 0, 0, 0.6);
-      }
-    }
-  }
+}
+:host {
+  display: block;
+  height: 100%;
+  overflow-y: auto; /* Enables scrolling */
 }
