+++ src/app/shared/components/blog-card/blog-card.component.html
@@ -12,7 +12,7 @@
                 class="card-image"
                 src="{{ entry.headerImageUrl }}"
                 alt="Blog Image"
-                onerror="this.src='https://media.istockphoto.com/id/1452662817/vector/no-picture-available-placeholder-thumbnail-icon-illustration-design.jpg?s=612x612&w=0&k=20&c=bGI_FngX0iexE3EBANPw9nbXkrJJA4-dcEJhCrP8qMw=';"
+                onerror="this.src='https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcRQU23E411UzSYZsdrhp-9gfWKUTqWkbtWVzA&s';"
         />
         <mat-card-content>
             <p>{{ entry.contentPreview }}</p>
