+++ src/app/shared/components/blog-card/blog-card.component.scss
@@ -1,0 +1,68 @@
+
+.blog-card {
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  box-sizing: border-box;
+  gap: 8px;
+
+  mat-card-header {
+    width: 100%;
+    max-width: 100%;
+    box-sizing: border-box;
+    display: flex;
+    flex-direction: column;
+    overflow: hidden;
+
+    .title {
+      white-space: nowrap;
+      overflow: hidden;
+      text-overflow: ellipsis;
+      display: block;
+      width: 100%;
+    }
+
+    .subtitle {
+      display: flex;
+      justify-content: space-between;
+      width: 100%;
+      overflow: hidden;
+    }
+  }
+
+  mat-card-image {
+    width: 100%;
+    height: 180px;
+    object-fit: cover;
+  }
+
+  mat-card-content {
+    flex-grow: 1;
+  }
+
+  mat-card-actions {
+    flex-shrink: 0;
+    display: flex;
+    justify-content: space-between;
+  }
+}
+
+.blog-card:hover {
+  box-shadow: 0 0 8px rgba(0, 0, 0, 0.1);
+  transform: translateY(-4px);
+  transition:
+          transform 0.2s,
+          box-shadow 0.2s;
+}
+
+.icon {
+  display: flex;
+  flex-direction: col;
+  align-items: center;
+  gap: 4px;
+}
+
+.icons {
+  display: flex;
+  gap: 8px;
+}
