+++ src/app/shared/components/blog-card/blog-card.component.ts
@@ -1,5 +1,5 @@
 import {Component, EventEmitter, Input, Output} from '@angular/core';
-import {BlogPreview} from '../../../features/models/blogPreview.model';
+import {BlogPreview} from '../../../core/models/blogPreview.model';
 import {MatCard, MatCardActions, MatCardContent, MatCardHeader, MatCardImage, MatCardSubtitle, MatCardTitle} from '@angular/material/card';
 import {MatIcon} from '@angular/material/icon';
 import {DatePipe} from '@angular/common';
