+++ src/main.ts
@@ -5,7 +5,7 @@ import {provideRouter} from '@angular/router';
 import {routes} from './app/app.routes';
 import {provideAnimationsAsync} from '@angular/platform-browser/animations/async';
 import {provideAuth} from 'angular-auth-oidc-client';
-import { authConfig } from './app/auth/auth.config';
+import { authConfig } from './app/core/auth/auth.config';
 
 bootstrapApplication(AppComponent, {
   providers: [
